tests: Move Windows tests to integration_windows.rs
Move the Windows guest test module from integration.rs into
its own standalone test binary integration_windows.rs. This
follows the same pattern used by integration_cvm.rs and
reduces the size of the main integration test file.

The new file includes all necessary imports and shares the
common test utilities via mod common.
